Add categorisation in utils
