+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@chatwoot/utils",
-  "version": "0.0.8",
+  "version": "0.0.9",
   "description": "Chatwoot utils",
   "private": false,
   "license": "MIT",
